tags working

+++ reimp/src/components/EditableIssueTag.js
@@ -8,7 +8,7 @@ import TagForm from './form/TagForm'
 import {
     updateIssueSubject,
     getIssues,
-    addTagToIssues,
+    addOrEditIssueTag,
     deleteTagFromIssues
 } from '../actions/Issues'
 import { updateTags, ensureTagsLoaded, getTag } from '../actions/Tags'
@@ -39,49 +39,49 @@ class EditableIssueTag extends Component {
     }
     
     onChange(new_value) {
-        const { dispatch } = this.props
-        dispatch(updateTags(new_value.tags))
+        const { dispatch, tag_id, issue_ids } = this.props
+        dispatch(addOrEditIssueTag(new_value.name, new_value.category_name, issue_ids, tag_id))
     }
 
-    onDelete(tag_id) {
-        const { dispatch, issue_ids } = this.props
+    onDelete() {
+        const { dispatch, issue_ids, tag_id } = this.props
+        if ( ! confirm("Delete this tag?") ) {
+            return false
+        }
         dispatch(deleteTagFromIssues(tag_id, issue_ids))
     }
 
-    onCreate(tag_name, tag_category_name) {
-        const { dispatch, issue_ids } = this.props
-        dispatch(addTagToIssues(tag_name, tag_category_name, issue_ids))
-    }
-
     render() {
-        const { issue, can_edit, tag, tag_id } = this.props
+        const { issue, can_edit, tag_id } = this.props
 
         return (
             <div>
-              { tag.id &&
+              { tag_id &&
                 <EditableProperty property_key={'issue_tag_'+tag_id}
                                   initial_value={tag_id}
                                   onChange={this.onChange}
-                                  can_edit={true}
+                                  can_edit={can_edit}
                     >
                   <TagForm tag_id={tag_id} />
-                  <Tag tag_id={tag_id} />
+                  <Tag tag_id={tag_id} can_edit={can_edit} />
                   <div className="text-component--empty"></div>
                 </EditableProperty>
               }
-              { tag.id &&
+              { tag_id && can_edit && 
                 <button className="button button--danger issue_sidebar--button" onClick={this.onDelete}>delete</button>
               }
-              { ! tag.id &&
+              { ! tag_id &&
                 <EditableProperty property_key={'issue_tag_new'}
                                   initial_value=''
                                   onChange={this.onChange}
-                                  can_edit={true}
+                                  can_edit={can_edit}
                     >
                   <TagForm tag_id={tag_id} />
                   <div className="text-component--readonly"></div>
                   <div className="text-component--empty">
-                    <button className="button button--primary issue_sidebar--button">Create tag</button>
+                    { can_edit && 
+                      <button className="button button--primary issue_sidebar--button">Create tag</button>
+                    }
                   </div>
                 </EditableProperty>
               }
@@ -100,6 +100,7 @@ function mapStateToProps(state, props) {
         tag,
         issue_ids,
         tag_id,
+        can_edit: true
     }
 }
